# backend/src/KapitelShelf.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "KapitelShelf.Api, Version=0.3.1.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Fetch all books.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Create a new book.",
        "requestBody": {
          "description": "The create book dto.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              }
            }
          }
        }
      }
    },
    "/books/search": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Search books with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        }
      }
    },
    "/books/search/suggestions": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Search suggestions with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/import": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Import a new book.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookFile": {
                    "type": "string",
                    "description": "Thebook file to import.",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "bookFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/books/import/asin": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Import a new book from its asin.",
        "parameters": [
          {
            "name": "asin",
            "in": "query",
            "description": "The asin of the book to import.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResultDTO"
                }
              }
            }
          }
        }
      }
    },
    "/books/{bookId}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get book by the id.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The id of the user.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to delete.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to update.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated book.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BookDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/books/{bookId}/cover": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get the cover for a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to get the cover for.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Add the cover for a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverFile": {
                    "type": "string",
                    "description": "The cover image file.",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "coverFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/books/{bookId}/file": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get the file for a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to get the file for.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Add the file for a book.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "description": "The id of the book to get.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookFile": {
                    "type": "string",
                    "description": "The book file.",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "bookFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search books with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The page number.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookDTOPagedResult"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/search/suggestions": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search suggestions with a search term.",
        "parameters": [
          {
            "name": "searchterm",
            "in": "query",
            "description": "The search term.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookDTO"
                  }
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the book id.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Gets or sets the title.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Gets or sets the description.",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "description": "Gets or sets the release date.",
            "format": "date-time",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "description": "Gets or sets the page number.",
            "format": "int32",
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/SeriesDTO"
          },
          "seriesNumber": {
            "type": "integer",
            "description": "Gets or sets the series number.",
            "format": "int32"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorDTO"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryDTO"
            },
            "description": "Gets or sets the categories.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            },
            "description": "Gets or sets the tags.",
            "nullable": true
          },
          "cover": {
            "$ref": "#/components/schemas/FileInfoDTO"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDTO"
          }
        },
        "additionalProperties": false,
        "description": "The book dto."
      },
      "BookDTOPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookDTO"
            },
            "description": "Gets or sets the items.",
            "nullable": true
          },
          "totalCount": {
            "type": "integer",
            "description": "Gets or sets the total number of items.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "The paginated result."
      },
...
      "CreateBookDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Gets or sets the title.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Gets or sets the description.",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "description": "Gets or sets the release date.",
            "format": "date-time",
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "description": "Gets or sets the page number.",
            "format": "int32",
            "nullable": true
          },
          "series": {
            "$ref": "#/components/schemas/CreateSeriesDTO"
          },
          "seriesNumber": {
            "type": "integer",
            "description": "Gets or sets the series number.",
            "format": "int32",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/CreateAuthorDTO"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCategoryDTO"
            },
            "description": "Gets or sets the categories.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTagDTO"
            },
            "description": "Gets or sets the tags.",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/CreateLocationDTO"
          }
        },
        "additionalProperties": false,
        "description": "The create dto for a book."
      },
...
      "CreateSeriesDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Gets or sets the name.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The create dto for a series."
      },
...
      "ImportResultDTO": {
        "type": "object",
        "properties": {
          "isBulkImport": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether it was a bulk import."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets or sets any errors encountered during the import process.",
            "nullable": true
          },
          "importedBooks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBookDTO"
            },
            "description": "Gets or sets the successfully imported books, containing their IDs and titles.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the result of an import operation, including details about the success or failure of the process."
      },
...
    }
  }
}